Recovery phase is implemented.
A recovery phase is implemented.
When a filesystem finds the node block with FSYNC_BIT in a roll-forward recovery phase, it checks if the node block is a master node as check its ofs.
If ofs is MUFIT_NODE_OFFSET, it read the node block as master node block.
And, it read block addresses in master node and recover them.
The master node block is truncated after roll-forward recovery or when the ioctl with F2FS_IOC_END_ATOMIC_FILES is called.

- Joontaek Oh