Fix some bugs
I fixed some code because of bugs.
First, if a file which in some atomic file set is closed, the atomic file set is released right a way.
In the original implementation, a sequence of closing file was random. So, the bugs can be occured.
Second, a offset of list_head in atomic_files_header has changed. The offset of list_head should be 0.
Third, a infinite loop in f2fs_do_sync_files() has resolved.
Because the files that is not last file in atomic file set does not have FSYNC_BIT, f2fs_do_sync_files() call
fsync_nodes_files() infinitely. This code should be called only in last file.

Joontaek Oh.