The directory "test_code" is updated.
The directory "test_code" is updated.
At the firs,t the subdirectory "microbenchmark" is added. The directory is for mataining experiment of microbenchmakr such as 4KB Random write followed by fsync().
And, some sciprts and results are added.
YCSB and RocksDB directories couldn't be added because their size.

Joontaek Oh.