Per-Group in-mem page list and hash table is added.
Per-Group in-mem page list and Hash table is added.
But, there are two problems.
First, when the per-group page list is implemented, it becomes slow.
Second, when the hash table is implemented, it becomes fast.
Becoming fast is good, but adding hash table is not for making fast but secure.
We have to figure them out.

Joontaek Oh.