Two problems I mentioned last commit are solved.
Two problems I mentioned last commit are solved.
For the first problem, it was because I didn't added code for dispatching requests made at the beginning of commit.
So I added code for dispatching requests just before loop of creating node requests.
For the second problem, it was because the f2fs_ioc_add_atomic_file has error, so there was only one file in the atomic file set.
I fixed it also.

Joontaek Oh.