Stealable transaction is implemented.
Stealable transaction is implemented. Transaction pages can be writebacked
to a storage even during a transaction.
When there are a few pages or a foreground gc is triggered, all transaction
pages are dirtied and writebacked to a storage.

Issue: I think there are only three cases when a foreground gc migrates
a transaction page; Not modified tx page, just modified tx page, and
a stolen page. In a case of stealed page, an old block address which is
in a victim segment has to be either an old block address which is a
block address before being stolen page or a new block address which is
a block address after being stolen page. However, there was a migrated
page whose both an old block address and a new block address are not
matched with an old block address which is in a victim segment.
It has to be checked.

Joontaek Oh.