No deadlock but operations fail.
No deadlock now.
It was because a bio of the file with FI_ATOMIC_FILE is treated as a F2FS_WB_CP_DATA.
However, it still has a problem.
In the YCSB Workload A for the rocksdb, at the run phase, a read and an update operations fail.
This problem also occurs even in the pure rocksdb.
We should examine why this problem occurs.

Joontaek Oh.